feat(dashboard): migrate routes from _layout to _authenticated
Replace the hand-rolled _layout.tsx sidebar with AuthenticatedLayout
(SidebarProvider + AppSidebar + SidebarInset, reading variant and
collapsible from useLayout()). Index and sandbox move over with their
actual content untouched - both now render their own <Header>/<Main>
instead of relying on a single shared wrapper, since Header/Main are
now reusable per-page components rather than layout-owned chrome.

main.tsx provider order: I18nProvider -> QueryClientProvider ->
ThemeProvider -> FontProvider -> LayoutProvider -> RouterProvider.

Verified in a real browser, not just the build: zero console errors
on / and /sandbox, active-state highlighting correct on both including
nested Sandbox, and the Auth collapsible group expands to its 5
sub-items with the chevron rotating as expected.
