+++ apps/dashboard/src/main.tsx
@@ -13,6 +13,8 @@ import { createRoot } from 'react-dom/client';
 import { RouterProvider, createRouter } from '@tanstack/react-router';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ThemeProvider } from './components/theme-provider';
+import { FontProvider } from './context/font-provider';
+import { LayoutProvider } from './context/layout-provider';
 import { I18nProvider } from '@repo/i18n';
 import './index.css';
 
@@ -48,7 +50,11 @@ createRoot(document.getElementById('root')!).render(
     >
       <QueryClientProvider client={queryClient}>
         <ThemeProvider defaultTheme="dark" attribute="class">
-          <RouterProvider router={router} />
+          <FontProvider>
+            <LayoutProvider>
+              <RouterProvider router={router} />
+            </LayoutProvider>
+          </FontProvider>
         </ThemeProvider>
       </QueryClientProvider>
     </I18nProvider>
