+++ apps/dashboard/src/routes/_authenticated/route.tsx
@@ -1,0 +1,6 @@
+import { createFileRoute } from '@tanstack/react-router';
+import { AuthenticatedLayout } from '../../components/layout/authenticated-layout';
+
+export const Route = createFileRoute('/_authenticated')({
+  component: AuthenticatedLayout,
+});
